Split out date validator
